# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Order from the Universe API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Universe order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order with the universe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Active order already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/active:
    get:
      operationId: getActiveOrder
      tags: [orders]
      summary: Get the current active order for a session
      parameters:
        - in: query
          name: sessionToken
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active order or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveOrderResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/checkout:
    post:
      operationId: createOrderCheckout
      tags: [orders]
      summary: Create a $1 Stripe checkout session for the order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCheckoutInput"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/confirm:
    patch:
      operationId: confirmOrderDelivery
      tags: [orders]
      summary: Confirm that the universe delivered what was ordered
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderConfirmInput"
      responses:
        "200":
          description: Order confirmed as delivered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /order-product:
    get:
      operationId: getOrderProduct
      tags: [orders]
      summary: Get the universe order product price info
      responses:
        "200":
          description: Product price info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderProduct"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/current:
    get:
      operationId: getCurrentOrders
      tags: [orders]
      summary: Get all active (non-delivered) orders for a session
      parameters:
        - in: query
          name: sessionToken
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of active orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/history:
    get:
      operationId: getOrderHistory
      tags: [orders]
      summary: Get all past delivered orders for a session
      parameters:
        - in: query
          name: sessionToken
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of past orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OrderStatus:
      type: string
      enum:
        - pending_payment
        - processing
        - in_transit
        - delivered

    Order:
      type: object
      required:
        - id
        - intention
        - status
        - sessionToken
        - motivationalMessage
        - trackingStage
        - createdAt
      properties:
        id:
          type: number
        intention:
          type: string
          description: What the person wants from the universe
        status:
          $ref: "#/components/schemas/OrderStatus"
        sessionToken:
          type: string
        stripePaymentIntentId:
          type: ["string", "null"]
        stripeCheckoutSessionId:
          type: ["string", "null"]
        motivationalMessage:
          type: string
          description: Uplifting message for this stage of the order
        trackingStage:
          type: number
          description: 1=pending, 2=processing, 3=in_transit, 4=delivered
        confirmedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    OrderInput:
      type: object
      required:
        - intention
        - sessionToken
      properties:
        intention:
          type: string
          minLength: 3
          maxLength: 500
          description: What you want to order from the universe
        sessionToken:
          type: string
          description: Client-generated session identifier

    OrderCheckoutInput:
      type: object
      required:
        - sessionToken
      properties:
        sessionToken:
          type: string

    OrderConfirmInput:
      type: object
      required:
        - sessionToken
      properties:
        sessionToken:
          type: string

    ActiveOrderResponse:
      type: object
      required:
        - order
      properties:
        order:
          oneOf:
            - $ref: "#/components/schemas/Order"
            - type: "null"

    CheckoutResponse:
      type: object
      required:
        - url
      properties:
        url:
          type: string

    OrderProduct:
      type: object
      required:
        - priceId
        - unitAmount
        - currency
      properties:
        priceId:
          type: string
        unitAmount:
          type: number
          description: Amount in cents (100 = $1.00)
        currency:
          type: string

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string